+++ testing/index.ts
@@ -1,3 +1,7 @@
 export { createMockRoom, linkMockRooms } from "./createMockRoom";
-export { openTwoPeers, captureConsoleErrors } from "./playwrightHelpers";
-export type { MeshPair } from "./playwrightHelpers";
+export {
+  openTwoPeers,
+  openNPeers,
+  captureConsoleErrors,
+} from "./playwrightHelpers";
+export type { MeshPair, MeshGroup } from "./playwrightHelpers";